requirements.txt
flask
flask-cors
pymongo[srv]
python-dotenv
elevenlabs
openai
google-genai